reorg data and bug fix

+++ high_dim_regression.py
@@ -4,14 +4,13 @@ from collections import defaultdict
 import json
 
 import numpy as np
-np.set_printoptions(precision=8, suppress=True)
 import pandas as pd
 
 from utils import eval_over_datasets
 from routines import run_lasso, run_rs_regression
 from data import isotropic_predictor_data
 
-
+np.set_printoptions(precision=8, suppress=True)
 parser = argparse.ArgumentParser()
 parser.add_argument('--num_samples', type=int, default=1000)
 parser.add_argument('--predictor_dim', type=int, default=100)
@@ -23,8 +22,6 @@ parser.add_argument('--lr', type=float, default=1e-2)
 parser.add_argument('--epoch', type=int, default=10000)
 parser.add_argument('--device', type=str, default='cuda:0')
 parser.add_argument('--num_alpha', type=int, default=5)
-
-
 
 if __name__ == "__main__":
     args = parser.parse_args()
@@ -44,26 +41,48 @@ if __name__ == "__main__":
     # run lasso if there is no lasso record
     lasso_file = os.path.join(output_folder, 'lars_metrics.csv')
     # re run the lasso
-    # if os.path.exists(lasso_file):
-        # print(f"lasso file {lasso_file} already found")
-        # lasso_df = pd.read_csv(lasso_file)
-    # else:
-    print(f"lasso file {lasso_file} not found")
-    data = defaultdict(list)
-    for alpha in alpha_range.tolist():
-        data['alpha'].append(alpha)
+    if os.path.exists(lasso_file):
+        print(f"lasso file {lasso_file} already found")
+        lasso_df = pd.read_csv(lasso_file)
+    else:
+        print(f"lasso file {lasso_file} not found")
+        data = defaultdict(list)
+        for alpha in alpha_range.tolist():
+            data['alpha'].append(alpha)
 
-        lasso_fetch = run_lasso(alpha, x, y, method='LARS')
-        data['time'].append(lasso_fetch['time'])
-        lasso_eval_fetch = eval_over_datasets(x, y, lasso_fetch['weights'], alpha)
-        for k in lasso_eval_fetch:
-            data[k].append(lasso_eval_fetch[k])
-    lasso_df = pd.DataFrame(data)
-    lasso_df.to_csv(lasso_file, index=False)
-    print(lasso_df.to_string())
+            lasso_fetch = run_lasso(alpha, x, y, method='LARS')
+            data['time'].append(lasso_fetch['time'])
+            lasso_eval_fetch = eval_over_datasets(x, y, lasso_fetch['weights'], alpha)
+            for k in lasso_eval_fetch:
+                data[k].append(lasso_eval_fetch[k])
+        lasso_df = pd.DataFrame(data)
+        lasso_df.to_csv(lasso_file, index=False)
+        print(lasso_df.to_string())
 
+    # run lasso if there is no lasso record
     lasso_file = os.path.join(output_folder, 'lasso_metrics.csv')
-    lasso_df = pd.read_csv(lasso_file)
+    # re run the lasso
+    if os.path.exists(lasso_file):
+        print(f"lasso file {lasso_file} already found")
+        lasso_df = pd.read_csv(lasso_file)
+    else:
+        print(f"lasso file {lasso_file} not found")
+        data = defaultdict(list)
+        for alpha in alpha_range.tolist():
+            data['alpha'].append(alpha)
+
+            lasso_fetch = run_lasso(alpha, x, y, method='lasso')
+            data['time'].append(lasso_fetch['time'])
+            lasso_eval_fetch = eval_over_datasets(x, y, lasso_fetch['weights'], alpha)
+            for k in lasso_eval_fetch:
+                data[k].append(lasso_eval_fetch[k])
+        lasso_df = pd.DataFrame(data)
+        lasso_df.to_csv(lasso_file, index=False)
+        print(lasso_df.to_string())
+
+        lasso_file = os.path.join(output_folder, 'lasso_metrics.csv')
+        lasso_df = pd.read_csv(lasso_file)
+
 
     # run rs
     data = defaultdict(list)
